requirements.txt
blowfish==0.6.1
freezegun==1.0.0
google-cloud-storage==1.42.0
gmplot==1.4.1
gps==3.19
grpcio-tools==1.41.1
guizero==1.1.1
haversine==2.3.0
numpy==1.19.4
obd==0.7.1
Pillow>=8.1.1
Pint==0.7.2
protobuf==3.14.0
pyppeteer==0.2.5
pyserial==3.5
python-dateutil==2.8.1
python-settings==0.2.2
pyttsx3==2.90
PyYAML>=5.4
six==1.15.0
websocket-client==0.57.0
